WIP. Refactoring (huge).
It seems that PocoClass brings too much complexity
and has too many weaknesses (not enough control
of the Poco space)... Or if they are really under control they
are too restricted. I'm about to totally remove them:
IPoco is the only workspace and other objects must be
converted from/to IPoco.