WIP: Back to basics.
- Removed any PocoClass idea.
- Attempt to Implement Covariance adapters.
It's complex... "Immutable Poco" will have to wait.

Back to Data Exchange aspect that is the priority.

Read only properties will ONLY drive CtorInstantiated properties
and may be, if they are nullable, "optional" properties.
The only possible "abstract" construction will be [CKTypeDefiner]IPoco
(IPocoPart) that may appear:
- as a writable or read only property.
- or in a generic parameter of a collection.

Focus is on soundness of the type system for the moment.
All collections MUST be concrete (no IReadOnlyXXX support) even
in the generic parameters.
For instance, a read only value tuple property with a non nullable IPoco
field must be CtoInstantiated and the IPoco must be a "concrete"
one (not a IPocoPart).
Or a non nullable string MUST have a [DefaultValue]...
Array is back (but cannot be on a read only property since
this will never be "covariant").