Introduced a bad idea: RegularAndNominalInfo
This nominal stuff for collection must be simpler. It should be
the "regular and nominal": the nominal collection should be
the regular one. The problem is that declaring a IList<T> will
trigger the regular List<T> support (even if it not used).

We may:
- Flag a regular has been created as a
nominal and un-flag it when actually requested.
- Expose this flag (ICollectionPocoType.IsOnlyNominal) and
let exchangers decide.
- Or transform this IsOnlyNominal to be non exchangeable "at the end".
But there is no "end": this would need to have a TypeSystem.Lock() (that
may be a good idea anyway).

Or we don't care of the "extra support" of the regular collection.

ay