Changed my mind: the Oblivious is now NON nullable.
Even if this is somehow not true from an implementation
point of view, this leads to simpler code and is in-line with how
 .Net considers the oblivious NRT mode.

One important simplification is that Oblivious types are all non
nullable, except the value types where both nullable and
non nullable are oblivious. But the nullable one can easily be
skipped and handled around the non nullable one by serializers.

However, this is currently less "future proof": the ".Net oblivious" leaks.
We now export "T" instead of "T?" type names, and "null" may appear:
the type name no longer carry the nullability information.
I may decide to unify this for value type: today "int?" or "Number?" is
exported for value type. May be a better option would be to state that
"Everything is potentially nullable in Json" and consider that
"Export/Import is NOT validation", that validation is an orthogonal
aspect.

For the moment, this is what it is. A "int?" or "int" may be handled or
the difference may be ignored: Validation should be done anyway...