Records work but name management must be changed.
1 - The pair Name/SimplifiedName is useless...
If a protocol needs to support aliases, it is simpler for it
to build a secondary name map and use it.

2 - Ambiguities must be explicitly managed. Currently
projected name clashes are not resolved to a single
unambiguous type. Writers don't care but Readers
have to handle this each time. It will be simpler and safer
if a Name has a single unambiguous oblivious target type.

And this is not easy :(.