Totally removed "SimpleName".
Mapping names is not as easy. To be sure to be able
to restore a "L(Number)" as a List<double>, this List<double>
MUST be registered. One solution would be to add a
TypeSystem.AddImplicitTypeMapping( t1, t2 ) that will
automatically create any type that rely on t2 for all type
that rely on t1.
Even if this is doable (thanks to the BackReference linked list),
this is not a simple task.