Removed MarshallableInProcessTypes.
Renamed IFrontAutoService into IEndpointAutoService
and enum IsFrontService to IsEndPointService everywhere.

IEndPointService are not  originally "marshalled" they must be
provided by the endpoints.
The idea is to introduce an indirection from a scope service, the
indirection being explictly set right after the call to CreateScope()
and to resolve all IsEndPointServices  through this indirection.
Not sure how well the Microsoft scope of the Conformant DI can handle
this resolution hook. This has to be tested first.

# Conflicts:
#	CK.StObj.Engine/CKType/CKTypeKind.cs
#	CK.StObj.Engine/CKType/Collector/CKTypeCollector.MultipleInterfaces.cs
#	Tests/CK.StObj.Engine.Tests/Service/BasicServiceLifeTimeTests.cs
#	Tests/CK.StObj.Engine.Tests/Service/FrontServiceTests.cs