Simplifying: no more function override at the EndpointUbiquitousInfo level.
OverriddenEndpointUbiquitousInfoServices mechanism and
EndpointUbiquitousInfo are somehow redundant.
OverriddenEndpointUbiquitousInfoServices is more complex (implies
the code generation) but is more explicit and more efficient.

But we should keep the override by value (what is clumsy and useless
is the override by functions) to allow fully external code to not be
limited by a Endpoint implementation.

Another question is about the copy...
Should we assume the mutabilty of the "current ubiquitous info" or
keep the current copy on write that isolates the modified data?

The current isolation is clearly safer... It's unfortunately overkill
in 99.99% of the use case (how many times will there be multiple
endpoint containers interactions in the same scope?) but I prefer
to stay on the safe side here.