Changed my mind...
Using the "dynamic" EndpointUbiquitousInfo will be easier.
There is no more specific handling at the EndpointDefinition level:
users of a Endpoint must use the EndpointUbiquitousInfo to
inject overridden instances (and only instances).
(Note: This forces an early evaluation of all the ubiquitous info and this
is not ideal. We'll see if we reintroduce a Func<ScopeData,object> or
Func<IServiceProvider,object> factory later).

There's no more OverriddenEndpointUbiquitousInfoServices and handling
of the ScopeData constructor parameter. The EndpointUbiquitousInfo
must now be code generated but its implementation is rather easy (chosen
implementation has been optimized for a small number of ubiquitous service
services). Moreovoer, it will handle IAutoService mapping directly: there
cannot be anymore difference with "static code" in the ConfigureServices
since everything must now go through the EndpointUbiquitousInfo. Of
course, user EndpointDefinition.ConfigureServices can still do whatever it
wants to override existing registrations: ConfigureServices will be called
with the ServiceCollection prefilled with the ubiquitous registrations).