Working.
There is one more thing to handle: a default value
provider for ubiquitous info services. This will
enable "front end point" to not be obliged to know
all the ubiquitous info services: the default value resolution
will be injected if no registration is found for the type.

This new IUbiquitousInfoDefault<T> is something like:
public interface IUbiquitousInfoDefault<T> where T : class
{
   T Default { get; }
}
Is this a Scope or a Singleton? It can be both, even if there's no
reason for it to be scoped: the default must be the same regardless
of any context. Should we use this default provider to define
a ubiquitous info rather than relying on the attribute boolean flag?
This implies to handle deferred "IsUbiquitous" flag for services (to
propagate to specialization). Rather than replacing the attribute flag,
this interface should better be a required complement of the flag.
And since its sole responsibility should be to expose the default, an
abstract class my be better.