Found a better way to publish ubiquitous infos.
This will be centralized on the EndpointTypeManager and
implemented as internal static fields so that generated
code can use it easily and efficiently. Dynamic code will
be able to exploit the EndpointTypeManager if needed.
This hides these informations and is better in-line with
a future "fully code generated DI" approach (that would
only allow global container to be configured by
IRealObject.ConfigureServices methods.

Removed last(?) trace of the dead DefaultEndpointContext.