WIP. "ProperType" way...
This is NOT enough. The best is to postpone
Readonly and Immutability support.
I keep the current restrictions but stop trying to
prepare the ground without a clear picture of the
destination...
I'll now just implement the covariant wrappers
for the Secondary and Abstract Poco (currently missing)
for the IPoco rooted collections.
Current restrictions will be kept.