Layer implementation (no tests).
This may be a good idea but doesn't really help serialization issues.
Default is that a PocoType is exchangeable by default. The opposite
would be a nightmare considering the numerous "small types" that
exist. Making a type not exchangeable should be application-wide,
not "protocol dependent".
Endpoint restrictions should be managed at a higher level than the
serialization one.