Fixed a "conceptual" bug.
"Exchangeable names" cannot be used to identify a
type since the serialization layer must have a unique,
common, view of the type names: the SerializableNames
are the only ones that must be used to interact with the
serialization.
Renamed engine services.
Added a central IPocoSerializationServiceEngine.GetSerializableIndex.

Should now handle the "runtime filters" for serialization for Import
and for export...