WIP...
For a cleaner and more versatile serializer, a real type mapping
capability is required.
Something like:
type.Map( Func<IPocoType,IPocoType> mapper )
It should be:
ts.Map( t, mapper ) as PocoType doesn't know its TypeSystem...
But the final type may not exist :(.
We need it at least for anonymous => anonymous.UnnamedRecord.
And this would also do IList => List and ISet => HashSet.
For dictionary, because of the lack of efficient generalization it is
less useful.