WIP. Will continue to try to keep the legacy/new.
KindDetector should now be obsolete... But some semantics
has changed. For instance, we'll try to allow [IsMultiple] on
abstract classes.

To handle the new ExternalTypes approach on the Engine
level and new RegisterCKTypeAttribute, we need to give
the aspects the capability to play with the EngineConfiguration
BEFORE analyzing types.

(Actually we may have designed a specific
API for type configuration but it is easier and more "opened" to
let the new EngineAspect.Initialize play with the EngineConfiguration
at its own risk....)