Basic level.
Is a LoopParameterTypeGroup missing?
This beast "merges" 2 (or more) sibling LoopParameterType
WHEN they appear simultaneously in a method parameters.
But we'll have to wait for the availability of all the new instances...

May be it could be better to suuport tuple as inputs? Or may be this is
completely stupid. Will think to this tomorrow.