Doesn't seem to be a great idea.
Handling AbstractHandlerType doesn't bring much to the table
and introduces complexities. They are not "real" HandlerTypes.
They should be [CKTypeDefiner] and their CurrentHandler doesn't
make sense. Their Callables must not have _args (only _parameters)
and they would be here to only act as a template for the Callable
of real HandlerType (bound to non "abstract" types). They may
be seen as an optimization but a rather useless one.

Giving up on the AbstractHandlerType, a ICachedType must now
expose the Members (and not only the DeclaredMembers that were
enough for our need so far), but this may be a good thing for the
future and makes the cache more "standard".